+++ src/stats/index.ts
@@ -1,3 +1,4 @@
 export * from './StatsDashboard'
 export * from './statistics'
+export * from './statsSelectors'
 export * from './types'